[+] Adicionado novos parametros as configurações de Banco e Cedente ACBrLibBoleto.
git-svn-id: svn://svn.code.sf.net/p/acbr/code/trunk2@24010 6e92efe7-b92a-0410-a9ec-f9e4e41bb3a6

+++ Projetos/ACBrLib/Fontes/Comum/ACBrObjectSerializer.pas
@@ -395,9 +395,9 @@ begin
         tkBool:
           AIni.WriteBool(ASessao, Propertie.Name, AValue.AsBoolean);
         tkEnumeration:
-          AIni.WriteInt64(ASessao, Propertie.Name, AValue.AsOrdinal);
+          AIni.WriteInteger(ASessao, Propertie.Name, AValue.AsOrdinal);
         tkInteger:
-          AIni.WriteInt64(ASessao, Propertie.Name, AValue.AsInteger);
+          AIni.WriteInteger(ASessao, Propertie.Name, AValue.AsInteger);
         tkInt64:
           AIni.WriteInt64(ASessao, Propertie.Name, AValue.AsInt64);
         tkWString,
